refactor(settings/ai-providers): extract DecoCreditsHero
Move TOP_UP_PRESETS, QuickTopUp, creditColorClass, and DecoCreditsHero
out of org-ai-providers.tsx into ai-providers/deco-credits-hero.tsx.
Remove now-orphaned imports (ToggleGroup, ToggleGroupItem, useQuery,
RefreshCw01) from the orchestrator file.
